refactor: 記事データ取得関数を別ファイルに切り出す
